Move LBFargateStackConfig to stack pkg
